feat: card types 추가
